requirements.txt
PySide6>=6.7
Pillow>=10.0
numpy>=1.26
ezdxf>=1.3
gdstk>=0.9
